requirements.txt
streamlit
SpeechRecognition
google-api-python-client
python-dotenv
pyaudio

